requirements.txt
cos_python_sdk_v5==1.9.24
matplotlib==3.7.1
numpy==1.22.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
pandas==1.4.3
Pillow==9.0.1
prompt_toolkit==3.0.30
PySide2==5.15.2.1
PySide6==6.5.0
PySide6_Addons==6.5.0
PySide6_Essentials==6.5.0
qt_material==2.14
torch==1.12.0
torchsummary==1.5.1
torchvision==0.13.0
unisms==0.0.7
